Relocate TheHive and DingTalk alerters
I moved these out of `alerters.py` but didn't put them into the
correct folder the first time around.
