+++ elastalert/alerters/thehive.py
